requirements.txt
python-docx
spacy
pandas
python -m spacy download en_core_web_sm